A tag is its number again, and FixTag holds the standard's as constants
Igor: the number is what logs, the wire and a consumer's own tags are written
in, and an enum asked for a cast at every one of them. FixField.Tag,
FixFinding.Tag, FixParseError.Tag and LengthDataPairs are int, as
QuickFIX/n's are, and generate.py writes FixTag as a static class of
constants. Every pattern and case reads as before, FixField.Decimal { Tag:
FixTag.OrderQty }; a tag outside the standard is 25005, with no cast. A
loaded check writes a tag as its plain number.

The casts both ways leave the code, the tests, the examples and the stands;
README, SKILL and the release notes say it; D139 records the reversal.

DotGram.Finance.Tests 4081 and DotGram.Finance.Fix44.Tests 613 pass.

+++ benchmarks/DotGram.Benchmarks/Stand.cs
@@ -504,7 +504,7 @@ static partial class Stand
 			var count = 0;
 
 			foreach (var field in fields)
-				count += (int)field.Tag;
+				count += field.Tag;
 
 			return count;
 		}
@@ -1626,7 +1626,7 @@ static partial class Stand
 			var count = 0;
 
 			foreach (var field in fields)
-				count += (int)field.Tag;
+				count += field.Tag;
 
 			return count;
 		}
